adding MapD3 as a separate component
