+++ frontend/its-our-call/src/components/MapD3.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
-import stateData from '../data/usa-states-dimensions';
-import districtData from '../data/usa-districts-dimensions';
 import * as d3 from 'd3';
 import * as d3Geo from 'd3-geo';
 import * as topojson from 'topojson';
 import usTopoJsonData from '../data/topojson-us';
 import { GeometryObject } from '../../node_modules/@types/topojson-specification';
+
+import './MapD3.css';
 
 // tslint:disable:no-console
 interface MapSVGProps {
@@ -22,14 +22,10 @@ interface MapD3State {
   defaultFill: string;
   defaultStroke: string;
   outerStroke: string;
-  viewBox: string;
   doneBuildingPaths: boolean;
   paths: JSX.Element[];
   currentQueue: number;
 }
-
-// const DIMENSIONS_KEY = 'dimensions';
-const VIEWBOX_KEY = 'viewBox';
 
 let lastMapId = 0;
 let newMapID = function() {
@@ -37,8 +33,10 @@ let newMapID = function() {
   return `map${lastMapId}`;
 };
 
-// const stateZoneKeys = Object.keys(stateData).filter((key: string) => { return key !== VIEWBOX_KEY; });
-// const districtZoneKeys = Object.keys(districtData).filter((key: string) => { return key !== VIEWBOX_KEY; });
+// tslint:disable-next-line:no-any
+let countryClicked = function(params: any) {
+  console.log(params);
+};
 
 export default class MapD3 extends React.Component<MapSVGProps, MapD3State> {
   uniqueID: string | undefined;
@@ -48,13 +46,11 @@ export default class MapD3 extends React.Component<MapSVGProps, MapD3State> {
     this.zoneStrokeColor = this.zoneStrokeColor.bind(this);
     this.stateClickHandler = this.stateClickHandler.bind(this);
     this.clickHandler = this.clickHandler.bind(this);
-    this.buildViewBox = this.buildViewBox.bind(this);
     this.createMap = this.createMap.bind(this);
     this.state = {
       defaultFill: props.defaultFill ? props.defaultFill : 'rgba(0,0,0,0)',
       defaultStroke: props.defaultStroke ? props.defaultStroke : 'rgba(0,0,0,0)',
       outerStroke: '#C9C9C9',
-      viewBox: this.buildViewBox(this.props.mapType === 'state'),
       doneBuildingPaths: false,
       paths: [],
       currentQueue: 0,
@@ -64,7 +60,6 @@ export default class MapD3 extends React.Component<MapSVGProps, MapD3State> {
   componentDidMount() {
     let currentQueue = this.state.currentQueue + 1;
     this.setState({
-      viewBox: this.buildViewBox(this.props.mapType === 'state'),
       paths: [],
       currentQueue,
       doneBuildingPaths: false
@@ -75,7 +70,6 @@ export default class MapD3 extends React.Component<MapSVGProps, MapD3State> {
   componentWillReceiveProps(props: MapSVGProps) {
     let currentQueue = this.state.currentQueue + 1;
     this.setState({
-      viewBox: this.buildViewBox(props.mapType === 'state'),
       paths: [],
       currentQueue,
       doneBuildingPaths: false,
@@ -83,13 +77,12 @@ export default class MapD3 extends React.Component<MapSVGProps, MapD3State> {
   }
 
   createMap() {
-    // width={500}
-    // height={500}
-    var width = 500;
+    var width = 938;
     var height = 500;
+
     var projection = d3Geo.geoAlbersUsa()
-    .scale(1000)
-    .translate([width / 2, height / 2]);
+      .scale(900) // 500
+      .translate([width / 2, height / 1.75]);
     var path = d3Geo.geoPath().projection(projection);
     var svg;
     if (this.uniqueID) {
@@ -97,10 +90,14 @@ export default class MapD3 extends React.Component<MapSVGProps, MapD3State> {
     } else {
       svg = d3.select('body').append('svg');
     }
-    svg.attr('width', width).attr('height', height);
-    // var svg = d3.select('body').append('svg')
-    //   .attr('width', width)
-    //   .attr('height', height);
+
+    svg.attr('preserveAspectRatio', 'xMidYMid')
+    .attr('viewBox', '0 0 ' + width + ' ' + height)
+    .attr('width', width)
+    .attr('height', height);
+    // .attr("width", m_width)
+    // .attr("height", m_width * height / width);
+
     // TODO: load the json externally like this:
     //  d3.json("/d/4090846/us.json", function(error, us) {
     var us = usTopoJsonData;
@@ -109,15 +106,18 @@ export default class MapD3 extends React.Component<MapSVGProps, MapD3State> {
       .attr('class', 'land')
       .attr('d', path);
 
-    let mesh = topojson.mesh(us, us.objects.counties, 
-        function(a: GeometryObject, b: GeometryObject) { 
-          // tslint:disable-next-line
-          return a !== b && !((a.id as number) / 1000 ^ (b.id as number) / 1000); 
-        });
-    svg.insert('path', '.graticule')
-      .datum(mesh)
-      .attr('class', 'county-boundary')
-      .attr('d', path);
+    // let mesh = topojson.mesh(us, us.objects.counties, 
+    //     function(a: GeometryObject, b: GeometryObject) { 
+    //       // tslint:disable-next-line
+    //       return a !== b && !((a.id as number) / 1000 ^ (b.id as number) / 1000); 
+    //     });
+    // svg.insert('path', '.graticule')
+    //   .datum(mesh)
+    //   .attr('class', 'county-boundary')
+    //   .attr('d', path);
+    console.log('us.objects.states');
+    console.log(us.objects.states);
+
     svg.insert('path', '.graticule')
       .datum(topojson.mesh(us, us.objects.states, 
         function(a: GeometryObject, b: GeometryObject) { 
@@ -125,19 +125,17 @@ export default class MapD3 extends React.Component<MapSVGProps, MapD3State> {
         }
       ))
       .attr('class', 'state-boundary')
-      .attr('d', path);
+      .attr('d', path).on('click', countryClicked);
       // });
-  }
-
-  buildViewBox(isStateMap: boolean): string {
-    // build zoom and pan adjustments here?
-    if (isStateMap) {
-      // we need to pad the width of the state map since we're adding state circle shapes
-      let stateDimensions = stateData[VIEWBOX_KEY].split(' ');
-      stateDimensions[2] = `${1100}`;
-      return stateDimensions.join(' ');
-    }
-    return districtData[VIEWBOX_KEY];
+      // make interactive paths
+      // svg.insert('path', '.graticule')
+    //   .data(us.objects.states).enter().append('d') topojson.mesh(us, us.objects.states, 
+    //     function(a: GeometryObject, b: GeometryObject) { 
+    //       return a !== b;
+    //     }
+    //   ))
+    //   .attr('class', 'state-boundary')
+    //   .attr('d', path).on('click', countryClicked);
   }
 
   zoneFillColor(zoneKey: string): string {
@@ -177,7 +175,7 @@ export default class MapD3 extends React.Component<MapSVGProps, MapD3State> {
 
   render() {
     return (
-      <div id={this.uniqueID} />
+      <div className="MapD3" id={this.uniqueID} />
     );
   }
 }
